+++ CHANGELOG.md
@@ -34,6 +34,7 @@ _These changes represent the move from v1 libraries to v2 and begin a new change
 - tooling: gulp tasks rewritten in TypeScript, updated to latest
 - tooling: build system rewritten as cli
 - common: extend renamed to assign
+- sp: client side pages breaking updates in method signatures
 
 ### Fixed